Move some to tutorial dir and create kanban component
