+++ apps/angular-app/src/app/kanban/kanban-main/kanban-main.component.html
@@ -1,0 +1,1 @@
+<p>kanban-main works!</p>
