+++ apps/angular-app/src/app/kanban/kanban-main/kanban-main.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-kanban-main',
+  standalone: true,
+  imports: [],
+  templateUrl: './kanban-main.component.html',
+  styleUrl: './kanban-main.component.scss'
+})
+export class KanbanMainComponent {
+
+}
