+++ apps/angular-app/src/app/tutorial/details/details.component.scss
