+++ apps/angular-app/src/app/tutorial/home/home.component.scss
