+++ apps/angular-app/src/app/tutorial/home/home.component.ts
@@ -2,8 +2,8 @@ import { Component, inject } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { HousingLocationComponent } from '../housing-location/housing-location.component';
-import { HousingLocation } from '../housinglocation';
-import { HousingService } from '../housing.service';
+import { HousingLocation } from '../../housinglocation';
+import { HousingService } from '../../housing.service';
 
 @Component({
   selector: 'app-home',
