+++ apps/angular-app/src/app/tutorial/housing-location/housing-location.component.scss
