+++ apps/angular-app/src/app/tutorial/housing-location/housing-location.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { HousingLocation } from '../housinglocation';
+import { HousingLocation } from '../../housinglocation';
 import { RouterLink } from '@angular/router';
 
 @Component({
